requirements.txt
fastapi
uvicorn
openai
auth0-python
python-dotenv
requests